--- apps/api/v2/src/modules/repositories/membership/membership.module.ts
@@ -1,10 +1,0 @@
-import { MembershipRepository } from "@/modules/repositories/membership/membership.repository";
-import { PrismaModule } from "@/modules/services/prisma/prisma.module";
-import { Module } from "@nestjs/common";
-
-@Module({
-  imports: [PrismaModule],
-  providers: [MembershipRepository],
-  exports: [MembershipRepository],
-})
-export class MembershipModule {}
