+++ apps/api/v2/src/modules/repositories/memberships/memberships.module.ts
@@ -1,0 +1,10 @@
+import { MembershipsRepository } from "@/modules/repositories/memberships/memberships.repository";
+import { PrismaModule } from "@/modules/services/prisma/prisma.module";
+import { Module } from "@nestjs/common";
+
+@Module({
+  imports: [PrismaModule],
+  providers: [MembershipsRepository],
+  exports: [MembershipsRepository],
+})
+export class MembershipsModule {}
