--- apps/api/v2/src/modules/repositories/user/user.module.ts
@@ -1,10 +1,0 @@
-import { UserRepository } from "@/modules/repositories/user/user.repository";
-import { PrismaModule } from "@/modules/services/prisma/prisma.module";
-import { Module } from "@nestjs/common";
-
-@Module({
-  imports: [PrismaModule],
-  providers: [UserRepository],
-  exports: [UserRepository],
-})
-export class UserModule {}
