+++ apps/api/v2/src/modules/repositories/users/users.module.ts
@@ -1,0 +1,10 @@
+import { UsersRepository } from "@/modules/repositories/users/users.repository";
+import { PrismaModule } from "@/modules/services/prisma/prisma.module";
+import { Module } from "@nestjs/common";
+
+@Module({
+  imports: [PrismaModule],
+  providers: [UsersRepository],
+  exports: [UsersRepository],
+})
+export class UsersModule {}
