+++ apps/api/v2/test/utils/withNextAuth.ts
@@ -1,10 +1,10 @@
 import { NextAuthStrategy } from "@/modules/auth/strategies";
-import { UserRepository } from "@/modules/repositories/user/user.repository";
+import { UsersRepository } from "@/modules/repositories/users/users.repository";
 import { TestingModuleBuilder } from "@nestjs/testing";
 import { NextAuthMockStrategy } from "test/mocks/next-auth-mock.strategy";
 
 export const withNextAuth = (email: string, module: TestingModuleBuilder) =>
   module.overrideProvider(NextAuthStrategy).useFactory({
-    factory: (userRepository: UserRepository) => new NextAuthMockStrategy(email, userRepository),
-    inject: [UserRepository],
+    factory: (userRepository: UsersRepository) => new NextAuthMockStrategy(email, userRepository),
+    inject: [UsersRepository],
   });
